[refactor] Build shared packages before server-database and next. [add] Seeder for `messages` and `tags` tables and relation `tags_messages` [refactor] T value for ObservableCache class [refactor] Now just reset database by test `login` and `register`

+++ src/database-server/src/reset-database.ts
@@ -16,15 +16,17 @@ const instance = new Sequelize(config);
 (async () => {
  try {
   await instance.authenticate();
-  await instance.query(
-   "delete from users; " +
-    "delete from contacts; " +
-    "delete from messages; " +
-    "delete from tags; " +
-    "delete from cron_jobs_messages; " +
-    "delete from sending_messages; " +
-    "delete from tags_messages;"
-  );
+  if (process.env.NODE_ENV !== "production") {
+   await instance.query(
+    "delete from users; " +
+     "delete from contacts; " +
+     "delete from messages; " +
+     "delete from tags; " +
+     "delete from cron_jobs_messages; " +
+     "delete from sending_messages; " +
+     "delete from tags_messages;"
+   );
+  }
   console.log("Deleted all data from database");
  } catch (e) {
   console.error(e);
